[MacOS] Add NSScrollView so NSTableView can scroll
