+++ Xamarin.Forms.Platform.MacOS/Renderers/ListViewRenderer.cs
@@ -9,11 +9,11 @@ using Xamarin.Forms.Internals;
 
 namespace Xamarin.Forms.Platform.MacOS
 {
-	public class ListViewRenderer : ViewRenderer<ListView, NSTableView>
+	public class ListViewRenderer : ViewRenderer<ListView, NSView>
 	{
 
 		bool _disposed;
-		const int DefaultRowHeight = 44;
+		public const int DefaultRowHeight = 44;
 		ListViewDataSource _dataSource;
 		bool _estimatedRowHeight;
 		IVisualElementRenderer _headerRenderer;
@@ -23,9 +23,24 @@ namespace Xamarin.Forms.Platform.MacOS
 		bool _shouldEstimateRowHeight = true;
 		IListViewController Controller => Element;
 		ITemplatedItemsView<Cell> TemplatedItemsView => Element;
+		NSTableView _table;
+
+		public virtual NSTableView CreateNSTableView(ListView list)
+		{
+			var table = new NSTableView { SelectionHighlightStyle = NSTableViewSelectionHighlightStyle.SourceList };
+
+			//this is needed .. can we go around it ?
+			table.AddColumn(new NSTableColumn(""));
+			//this line hides the header by default
+			table.HeaderView = null;
+
+			table.Source = _dataSource = list.HasUnevenRows ? new UnevenListViewDataSource(list, _table) : new ListViewDataSource(list, _table);
+
+			return table;
+		}
+
 		protected override void Dispose(bool disposing)
 		{
-			// check inset tracker for null to 
 			if (disposing && !_disposed)
 			{
 				_disposed = true;
@@ -34,14 +49,14 @@ namespace Xamarin.Forms.Platform.MacOS
 				while (viewsToLookAt.Count > 0)
 				{
 					var view = viewsToLookAt.Pop();
-					//var viewCellRenderer = view as ViewCellRenderer.ViewTableCell;
-					//if (viewCellRenderer != null)
-					//	viewCellRenderer.Dispose();
-					//else
-					//{
-					//	foreach (var child in view.Subviews)
-					//		viewsToLookAt.Push(child);
-					//}
+					var viewCellRenderer = view as ViewCellNSView;
+					if (viewCellRenderer != null)
+						viewCellRenderer.Dispose();
+					else
+					{
+						foreach (var child in view.Subviews)
+							viewsToLookAt.Push(child);
+					}
 				}
 
 				if (Element != null)
@@ -72,7 +87,7 @@ namespace Xamarin.Forms.Platform.MacOS
 				var headerView = Controller?.HeaderElement as VisualElement;
 				if (headerView != null)
 					headerView.MeasureInvalidated -= OnHeaderMeasureInvalidated;
-				Control?.HeaderView?.Dispose();
+				_table?.HeaderView?.Dispose();
 
 				//var footerView = Controller?.FooterElement as VisualElement;
 				//if (footerView != null)
@@ -109,11 +124,12 @@ namespace Xamarin.Forms.Platform.MacOS
 			{
 				if (Control == null)
 				{
-					SetNativeControl(new NSTableView { SelectionHighlightStyle = NSTableViewSelectionHighlightStyle.SourceList });
-					//this is needed .. can we go around it ?
-					Control.AddColumn(new NSTableColumn("Values"));
-					Control.Source = _dataSource = e.NewElement.HasUnevenRows ? new UnevenListViewDataSource(e.NewElement, Control) : new ListViewDataSource(e.NewElement, Control);
-
+					var scroller = new NSScrollView
+					{
+						AutoresizingMask = NSViewResizingMask.HeightSizable | NSViewResizingMask.WidthSizable,
+						DocumentView = _table = CreateNSTableView(e.NewElement)
+					};
+					SetNativeControl(scroller);
 				}
 
 				_shouldEstimateRowHeight = true;
@@ -152,7 +168,7 @@ namespace Xamarin.Forms.Platform.MacOS
 			else if (e.PropertyName == ListView.HasUnevenRowsProperty.PropertyName)
 			{
 				_estimatedRowHeight = false;
-				Control.Source = _dataSource = Element.HasUnevenRows ? new UnevenListViewDataSource(_dataSource) : new ListViewDataSource(_dataSource);
+				_table.Source = _dataSource = Element.HasUnevenRows ? new UnevenListViewDataSource(_dataSource) : new ListViewDataSource(_dataSource);
 			}
 			else if (e.PropertyName == ListView.IsPullToRefreshEnabledProperty.PropertyName)
 				UpdatePullToRefreshEnabled();
@@ -216,7 +232,7 @@ namespace Xamarin.Forms.Platform.MacOS
 			var request = headerView.Measure(width, double.PositiveInfinity, MeasureFlags.IncludeMargins);
 			Xamarin.Forms.Layout.LayoutChildIntoBoundingRegion(headerView, new Rectangle(0, 0, width, request.Request.Height));
 
-			Control.HeaderView = (NSTableHeaderView)_headerRenderer.NativeView;
+			_table.HeaderView = (NSTableHeaderView)_headerRenderer.NativeView;
 		}
 
 		void OnScrollToRequested(object sender, ScrollToRequestedEventArgs e)
@@ -250,7 +266,7 @@ namespace Xamarin.Forms.Platform.MacOS
 						_headerRenderer.SetElement(headerView);
 						return;
 					}
-					Control.HeaderView = null;
+					_table.HeaderView = null;
 					var platform = _headerRenderer.Element.Platform as Platform;
 					if (platform != null)
 						platform.DisposeModelAndChildrenRenderers(_headerRenderer.Element);
@@ -270,7 +286,7 @@ namespace Xamarin.Forms.Platform.MacOS
 			}
 			else if (_headerRenderer != null)
 			{
-				Control.HeaderView = null;
+				_table.HeaderView = null;
 				var platform = _headerRenderer.Element.Platform as Platform;
 				if (platform != null)
 					platform.DisposeModelAndChildrenRenderers(_headerRenderer.Element);
@@ -282,7 +298,7 @@ namespace Xamarin.Forms.Platform.MacOS
 		void UpdateIsRefreshing()
 		{
 			var refreshing = Element.IsRefreshing;
-
+			System.Diagnostics.Debug.WriteLine("Is Refreshing not implemented");
 		}
 
 		void UpdateItems(NotifyCollectionChangedEventArgs e, int section, bool resetWhenGrouped)
@@ -372,10 +388,10 @@ namespace Xamarin.Forms.Platform.MacOS
 			var rowHeight = Element.RowHeight;
 			if (Element.HasUnevenRows && rowHeight == -1)
 			{
-				Control.RowHeight = NoIntrinsicMetric;
+				_table.RowHeight = NoIntrinsicMetric;
 			}
 			else
-				Control.RowHeight = rowHeight <= 0 ? DefaultRowHeight : rowHeight;
+				_table.RowHeight = rowHeight <= 0 ? DefaultRowHeight : rowHeight;
 		}
 
 		void UpdateSeparatorColor()
@@ -406,211 +422,9 @@ namespace Xamarin.Forms.Platform.MacOS
 			System.Diagnostics.Debug.WriteLine("Separator not implemented on MacOS");
 		}
 
-		internal class UnevenListViewDataSource : ListViewDataSource
-		{
-			IVisualElementRenderer _prototype;
 
 
-			public UnevenListViewDataSource(ListView list, NSTableView tableView) : base(list, tableView)
-			{
-			}
 
-			public UnevenListViewDataSource(ListViewDataSource source) : base(source)
-			{
-			}
-
-			internal nfloat CalculateHeightForCell(NSTableView tableView, Cell cell)
-			{
-				var viewCell = cell as ViewCell;
-				if (viewCell != null && viewCell.View != null)
-				{
-					var target = viewCell.View;
-					if (_prototype == null)
-					{
-						_prototype = Platform.CreateRenderer(target);
-						Platform.SetRenderer(target, _prototype);
-					}
-					else
-					{
-						_prototype.SetElement(target);
-						Platform.SetRenderer(target, _prototype);
-					}
-
-					var req = target.Measure(tableView.Frame.Width, double.PositiveInfinity, MeasureFlags.IncludeMargins);
-
-					target.ClearValue(Platform.RendererProperty);
-					foreach (var descendant in target.Descendants())
-						descendant.ClearValue(Platform.RendererProperty);
-
-					return (nfloat)req.Request.Height;
-				}
-				var renderHeight = cell.RenderHeight;
-				return renderHeight > 0 ? (nfloat)renderHeight : DefaultRowHeight;
-			}
-		}
-
-		internal class ListViewDataSource : NSTableViewSource
-		{
-			const int DefaultItemTemplateId = 1;
-			static int s_dataTemplateIncrementer = 2; // lets start at not 0 because
-			readonly nfloat _defaultSectionHeight;
-			readonly Dictionary<DataTemplate, int> _templateToId = new Dictionary<DataTemplate, int>();
-			readonly NSTableView _uiTableView;
-			protected readonly ListView List;
-			IListViewController Controller => List;
-			ITemplatedItemsView<Cell> TemplatedItemsView => List;
-			bool _isDragging;
-			bool _selectionFromNative;
-
-			public ListViewDataSource(ListViewDataSource source)
-			{
-				List = source.List;
-				_uiTableView = source._uiTableView;
-				_defaultSectionHeight = source._defaultSectionHeight;
-				_selectionFromNative = source._selectionFromNative;
-
-				Counts = new Dictionary<int, int>();
-			}
-
-			public ListViewDataSource(ListView list, NSTableView tableView)
-			{
-				_uiTableView = tableView;
-
-				List = list;
-				List.ItemSelected += OnItemSelected;
-				UpdateShortNameListener();
-
-				Counts = new Dictionary<int, int>();
-			}
-
-			public Dictionary<int, int> Counts { get; set; }
-
-
-			public void OnItemSelected(object sender, SelectedItemChangedEventArgs eventArg)
-			{
-				if (_selectionFromNative)
-				{
-					_selectionFromNative = false;
-					return;
-				}
-
-				var location = TemplatedItemsView.TemplatedItems.GetGroupAndIndexOfItem(eventArg.SelectedItem);
-				if (location.Item1 == -1 || location.Item2 == -1)
-				{
-					//	var selectedIndexPath = _uiTableView.IndexPathForSelectedRow;
-
-					//	var animate = true;
-
-					//	if (selectedIndexPath != null)
-					//	{
-					//		var cell = _uiTableView.CellAt(selectedIndexPath) as ContextActionsCell;
-					//		if (cell != null)
-					//		{
-					//			cell.PrepareForDeselect();
-					//			if (cell.IsOpen)
-					//				animate = false;
-					//		}
-					//	}
-
-					//	if (selectedIndexPath != null)
-					//		_uiTableView.DeselectRow(selectedIndexPath, animate);
-					return;
-				}
-
-				//_uiTableView.SelectRow(NSIndexPath.FromRowSection(location.Item2, location.Item1), true, UITableViewScrollPosition.Middle);
-			}
-
-			public override nint GetRowCount(NSTableView tableView)
-			{
-				int section = 1;
-				int countOverride;
-				if (Counts.TryGetValue((int)section, out countOverride))
-				{
-					Counts.Remove((int)section);
-					return countOverride;
-				}
-
-				var templatedItems = TemplatedItemsView.TemplatedItems;
-				if (List.IsGroupingEnabled)
-				{
-					var group = (IList)((IList)templatedItems)[(int)section];
-					return group.Count;
-				}
-
-				return templatedItems.Count;
-			}
-
-			const string identifer = "myCellIdentifier";
-
-			public override NSView GetViewForItem(NSTableView tableView, NSTableColumn tableColumn, nint row)
-			{
-				var indexPath = NSIndexPath.FromItemSection(row, 1);
-				var id = TemplateIdForPath(indexPath);
-				var cell = GetCellForPath(indexPath);
-				var nativeCell = CellNSView.GetNativeCell(tableView, cell, true, id.ToString());
-				return nativeCell;
-			}
-
-			public void UpdateGrouping()
-			{
-				UpdateShortNameListener();
-				_uiTableView.ReloadData();
-			}
-
-			protected Cell GetCellForPath(NSIndexPath indexPath)
-			{
-				var templatedItems = TemplatedItemsView.TemplatedItems;
-				if (List.IsGroupingEnabled)
-					templatedItems = (TemplatedItemsList<ItemsView<Cell>, Cell>)((IList)templatedItems)[(int)indexPath.Section];
-
-				var cell = templatedItems[(int)indexPath.Item];
-				return cell;
-			}
-
-			void OnShortNamesCollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
-			{
-				_uiTableView.ReloadData();
-			}
-
-			int TemplateIdForPath(NSIndexPath indexPath)
-			{
-				var itemTemplate = List.ItemTemplate;
-				var selector = itemTemplate as DataTemplateSelector;
-				if (selector == null)
-					return DefaultItemTemplateId;
-
-				var templatedList = TemplatedItemsView.TemplatedItems;
-				if (List.IsGroupingEnabled)
-					templatedList = (TemplatedItemsList<ItemsView<Cell>, Cell>)((IList)templatedList)[(int)indexPath.Section];
-
-				var item = templatedList.ListProxy[(int)indexPath.Item];
-
-				itemTemplate = selector.SelectTemplate(item, List);
-				int key;
-				if (!_templateToId.TryGetValue(itemTemplate, out key))
-				{
-					s_dataTemplateIncrementer++;
-					key = s_dataTemplateIncrementer;
-					_templateToId[itemTemplate] = key;
-				}
-				return key;
-			}
-
-			void UpdateShortNameListener()
-			{
-				var templatedList = TemplatedItemsView.TemplatedItems;
-				if (List.IsGroupingEnabled)
-				{
-					if (templatedList.ShortNames != null)
-						((INotifyCollectionChanged)templatedList.ShortNames).CollectionChanged += OnShortNamesCollectionChanged;
-				}
-				else
-				{
-					if (templatedList.ShortNames != null)
-						((INotifyCollectionChanged)templatedList.ShortNames).CollectionChanged -= OnShortNamesCollectionChanged;
-				}
-			}
-		}
 	}
 
 }
